Nuevo mensaje corregido

+++ src/main/java/cl/duoc/ferremasapp/controller/UsuarioController.java
@@ -15,7 +15,7 @@ public class UsuarioController {
     private UsuarioService usuarioService;
 
     @GetMapping("/{id}")
-    public Optional<Usuario> obtenerUsuario(@PathVariable Long id) {
+    public Optional<Usuario> obtenerUsuario(@PathVariable Integer id) {
         return usuarioService.buscarPorId(id);
     }
 }
