+++ src/main/java/cl/duoc/ferremasapp/service/ProductoService.java
@@ -7,9 +7,9 @@ import cl.duoc.ferremasapp.model.Producto;
 
 public interface ProductoService {
     Producto registrarProducto(Producto producto);
-    Optional<Producto> buscarPorId(Long id);
+    Optional<Producto> buscarPorId(Integer id);
     Optional<Producto> buscarPorCodigo(String codigo); // Agregado
     List<Producto> listarProductos();  // Este es como findAll()
     Producto actualizarProducto(Producto producto);
-    void eliminarProducto(Long id);
+    void eliminarProducto(Integer id);
 }
